+++ Source/FicsitWiremod/Public/Behaviour/Gates/Numbers/LerpNumber.h
@@ -12,7 +12,7 @@ class FICSITWIREMOD_API ALerpNumber : public AFGWiremodBuildable
 	GENERATED_BODY()
 
 public:
-	virtual void Process_Implementation(float DeltaTime) override
+	virtual void Process_Implementation(double DeltaTime) override
 	{
 		auto A = GetConnection(0).GetFloat();
 		auto B = GetConnection(1).GetFloat();
@@ -30,5 +30,5 @@ public:
 	}
 
 	UPROPERTY(Replicated, SaveGame)
-	float Out;
+	double Out;
 };
