+++ Source/FicsitWiremod/Public/Behaviour/Miscellaneous/Notificator.h
@@ -12,7 +12,7 @@ class FICSITWIREMOD_API ANotificator : public AFGWiremodBuildable
 	GENERATED_BODY()
 
 public:
-	virtual void Process_Implementation(float DeltaTime) override
+	virtual void Process_Implementation(double DeltaTime) override
 	{
 		auto Fire = GetConnection(0).GetBool();
 
@@ -31,7 +31,7 @@ public:
 	}
 
 	UFUNCTION(BlueprintImplementableEvent)
-	void ShowNotification(const FString& Title, const FString& Description, float Time);
+	void ShowNotification(const FString& Title, const FString& Description, double Time);
